refactor(ui): l'app principale est en tsx
